requirements.txt
playwright==1.42.0
ics==0.7.2
pytz==2024.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0
